RAG/requirements.txt
pypdf==4.0.1
langchain>=0.2.0,<0.3
langchain-community>=0.0.38
sentence-transformers==3.0.0
faiss-cpu==1.13.2
python-dotenv==1.0.0
google-generativeai==0.3.0
serpapi>=0.1.5
requests==2.31.0
beautifulsoup4==4.12.2